Support loading of aliased chunks/modules from external build.
At present, it is only possible for the Webpack runtime loader to
load chunks that were generated as part of the same build as the
entry point.

However, for distributed teams like the Walmart.com team, it is
advantageous to share chunks of code between different entry points;
both those loaded in the same browser context and otherwise.  The
primary benefit is client-side caching of code.  If 30 of our
sub-projects bundle Underscore.js, the chunk containing that
library would ideally be cached and re-used by each of those
sub-projects.

I first attempted to support this use case as a plugin.  However,
plugins for generating the entry bootstrap are applied in waterfall.
https://github.com/divmain/webpack/blob/dc81f2ed975b0bf49fbf4ca600222e1c071c0680/lib/MainTemplate.js#L101-L101

This fundamentally interfered and necessitated overriding templating
at a much higher level.  This is undesirable, as the plug-in could
quickly fall behind changes made to Webpack.

This commit provides the most basic functionality to define a
mapping of external to local chunks and modules.  The `externals`
object would map the external build namespace to local aliases for
the chunks and modules.

Example:

  {
    "other-build-namespace": {
      "chunks": {
        4: 7 // where 4 is the chunk ID from other-build and 7 is from this build
      },
      "modules": {
        5: 9 // where 5 is the module ID from other-build and 9 is from this build
      }
    }
  }

The onus would be on plugin authors to generate this externals object.

My plan is to build a plugin that aliases chunks from other Webpack
builds to their local counterparts at build-time.  These chunks will
be identical in every way except for the chunk and module IDs.  These
IDs will be mapped and made available as `externals` by exporting
asset manifests from consumable projects, and importing and processing
those manifests from consuming projects.