Small refactor for getting size
Currently, map is being called to extract sizes then reduce is called on that; so 2 iterations through the array. It is possible to solve this with only reduce: 1 iteration. reallyUsedModules probably won't ever be big enough for this to matter, so it comes down to readability.

I think doing a once-and-done reduce could be an improvement here. What do you guys think?