[WIP] Deduped modules should return the same exports
I have changed the dedupe algorithm to return the same exports as the deduped module. This ensures that required modules are identical across package boundaries.

This is really useful in the case of Angular2's dependency injector which uses Classes as Tokens (https://angular.io/docs/js/latest/api/di/Injector-class.html) 
Without this fix, the classes that are `required` are not identical across packages, so the DI Injector treats them as different Tokens.

I'm not sure if I should also update the `plugin("add-module")` section.