Name AMD module in UMD target if library is set
In the AMD target the AMD module gets named, but in the UMD target its
is always anonymous, even if `library` is set in the config.

This patch names the AMD module if `library` is set.

See the discussion in the referenced issue to see the background for
this change.

If its a multi part library it uses the last key in the same was as it
does with the exports hash part of the UMD declaration. Is this the best
solution?

Resolves webpack/issues/989