Use `interop-require` for loading configs.
With the release of `babel@6` the way generated modules work now does NOT set the `module.exports` property; instead it sets `exports.default`. Of course this basically breaks everything trying to `require()` it. This behavior is not changing: https://github.com/babel/babel/issues/2212.

So when using `interpret` and `babel@6` to have your configuration files in ES6 this shim is necessary.