Fixing issue where DllPlugin + code-splits caused Webpack to end without emitting chunks.
When utilizing code-splitting via a statement like `require.ensure` in addition to the `DllPlugin`, the webpack process would emit the Dll manifest, but not the chunks corresponding to that manifest.

According to [async#forEach](https://github.com/caolan/async#eacharr-iterator-callback), the callback must be called at the completion of every iteration - when utilizing code-splitting, the shortcut branch here would never call the iterator callback, which in turn meant that the `emit` callback never got called, causing the chunks generated by the Dll to not get emitted.

Adding the completion callback results in the behavior we expect - the manifest and the chunks corresponding to that manifest both get emitted.