Fix Windows filename backslash incompatibility
With some loaders (specifically with babel) under Windows the file names
in the source map already come converted to forward slashes. This is
incompatible with RequestShortener so that the source map paths are not
shortened to a relative path. This change makes RequestShortener tolerant
to such cases while preserving existing behaviour.

Fixes #1770