Add a lookup index for modules to Chunks.
Since Chunks often contain many many modules in large projects, being able to determine if a module is part of a chunk efficiently is important. This patch adds a lookup index off of the debugId which users of the chunk can use to determine if a module is in a chunk.

This patch also teaches the RemoveParentModulePlugin to use the index. In my profiling, that particular lookup is responsible for 99% of the time spent in the plugin, which is responsible for 95% of the rebuild time in large projects.

I did not apply this technique elsewhere, but it could be useful in many other places across the codebase. Adding a new `contains` method to collections that uses the index if available would make future changes easier to apply.