Remove child asset hashes from compilation hash compilation
Now that "extract-text-webpack-plugin" correctly adds asset hashes to compilation hash, this can be removed.

I'm actually not sure whether it makes sense to keep it (not aware of any other plugins affected by this),
But I'd remove it at least for now, because of how extract-text-webpack-plugin works today.
Child compilation is only added if css changes, which breaks test case watchCases/plugins/extract-text-plugin step 3 (no-op JS change) on Windows and Linux.
It works on macOS though, but only as a result of another bug: watchpack reports all files as changed on that platform.