Use hash of loader to determine its "ident" if its missing
As loader/plugin options as part of the webpack config are only
allowed as part of the options a lot of people forget to add an "ident"
when they migrate their loaders. This ident is requried as the options
get stringified along the way and therefore a "behind the back" lookup is
made with the ident to retrieve non stringifiable data.
This should be done automagically though and not left for the user unless
they want to specify an "ident" (imho).

this pr should fix issues that are currently arrising in css-loaders et all.