Reliable conditions for multiple provide plugins
If multiple provide plugins were used, excludes and includes would not
work reliably because we were using a single parserOptions globally to
control the behaviour of all of them. Fixed by assigning a UUID
identifying a plugin instance and keeping track of those when
instantiating the parser for a module.