Split sourceMappingURL string
This change is splitting `sourceMappingURL` to `source" + "MappingURL` in order to avoid syntax errors caused by babel when importing webpack in code that is transpiled by babel. This workaround was suggested in https://github.com/babel/babel/issues/4012 and this issue is an upstream to https://github.com/electron/electron-compilers/issues/67 and https://github.com/electron/electron-compile/issues/210 .

Apart from a small syntax change there are no changes to the functionality.