[performance] Speed up Module.rewriteChunkInReasons.
This speeds up the implementation of rewriteChunkInReasons. Previously this was
calling indexOf via addToSet(), inside of a loop (looping over newChunks),
inside of another loop (this.reasons). Essentially a triply nested loop, which
was killing performance.

The new implementation moves the loop over newChunks outside of the
this.reasons loop, and avoids the indexOf call by using an object to ensure
unique chunks.

On a very large build, this saw improvements at > 40 seconds.