fix: cli --watch-poll not accept number correctly
The document say "--watch-poll" option can determine "The polling
interval for watching (also enable polling)". But the previous
implement only accept boolean value and no way to set the interval.

This patch try to fix this problem by allow number input like this:
"--watch-poll=100".

This patch also try to keep the original behavior not be changed as much
as possible. So "--watch-poll" and "--watch-poll=true" are still
acceptable (enable the polling ability).

But still has one incompatible change: original "--watch-poll=" mean
disable polling but after this fix it mean enable polling. It's due to
yargs can not differentiate the different between "--watch-poll" and
"--watch-poll=" when using string type option. So no way to workaround
for this problem.