Introduce webpack/__boilerplate__/ prefix for ECMAScript module identifier
When an ECMAScript module identifier is prefixed with
webpack/__boilerplate__/ for importing, it would import the module but NOT
mark the importing module is ECMAScript module.
It is convenient when adding implicit import with loader or plugin. In such
a case, this change allows the module to utilize CommonJS features unless
it has explicit Harmony import or export declaration.