Reduce the number of calls to hash.update
I was profiling our webpack build on Node 8.9.0 and noticed that a fair
amount of time was spent calling hash.update. During the process of
building one of our bundles, ~115ms was spent inside of createHash. 95ms
of that time is spent in crypto's update itself. In the bottom up view
of the profiler, this is the 4th most expensive part of the profile, as
measured by self time.

It turns out that, concatenating up a string and calling crypto's update
once is much faster than calling update many times. In this commit, I
take a swing at replacing the pattern of passing around a hash object
that update is called on all the time, to concatenating up a string, and
reducing the number of calls to crypto's update.

After this change, my profiling shows that only 16.54ms is spent in
createHash. In the bottom up view, I verified that 8.5ms is now spent in
crypto's update, which moves it far from one of the most expensive
areas.

I think there are more opportunities like this, but I'm a little
hesitant to dig in much more at this point because of the staggering
amount of indirection which makes the code difficult to follow, the
profiling difficult to interpret, and the stack traces when I hit an
error not very helpful.