Allow jsonp promise global to be configured in output.promiseGlobal
This fixes the fact that presently, `new Promise` is hard-coded in
JsonpMainTemplatePlugin. This prevents using jsonp or chunking
without first polyfilling window.Promise. Polyfilling is not
advisable when building scripts to run on third party domains,
such as SDKs or libraries which need to be run across many
different sites where the browser environment is not under the
control of the script author.