Fix template tag syntax
In 2.9 and earlier, Typescript allowed braces around the name introduced
by the `@template` tag, even though this syntax is non-standard. In 3.0,
Typescript introduces a more complex `@template` syntax that uses braces
for specifying the constraint of a type parameter:

```js
/** @template {Constraint} T -- By analogy with param syntax:
 *  @param {T} name
 */
function f(name) { ... }
```

That means that the syntax `/** @template {T} */` is illegal. This PR
fixes the uses of the `@template` tag ahead of Typescript 3.0 as well as
typescript@next.