Catch errors while parsing DLL manifest
If the DLL manifest is blank or malformed, parsing it will throw
an error, and if left uncaught will kill the process if webpack is
in watch mode. Catching the error and adding it to the compilation's
errors will result in a failed compilation, but will keep the
process running while in watch mode.