Fix build crash due to webpack-command pathing
When running webpack in certain environments, one can get the following exception thrown:

```
path.js:39
    throw new ERR_INVALID_ARG_TYPE('path', 'string', path);
    ^

TypeError [ERR_INVALID_ARG_TYPE]: The "path" argument must be of type string. Received type undefined
    at assertPath (path.js:39:11)
    at Object.resolve (path.js:168:7)
    at Object.<anonymous> (D:\dev\SCAFFOLDING-webpack\node_modules\webpack\bin\webpack.js:167:15)
    at Module._compile (internal/modules/cjs/loader.js:702:30)
    at Object.Module._extensions..js (internal/modules/cjs/loader.js:713:10)
    at Module.load (internal/modules/cjs/loader.js:612:32)
    at tryModuleLoad (internal/modules/cjs/loader.js:551:12)
    at Function.Module._load (internal/modules/cjs/loader.js:543:3)
    at Function.Module.runMain (internal/modules/cjs/loader.js:744:10)
    at startup (internal/bootstrap/node.js:238:19)
```

due to the fact that `installedClis[0].binName` may resolve to `webpack-command` when the package directory is expected to be just `webpack`. When encountering `webpack-command` we adjust the value to `webpack` to avoid crashing and continue as expected.