fix records revival of used chunk and module ids
When running webpack in watch mode or using portable records, revival of chunk and module ids happens correctly but after it `usedChunkIds` and `usedModuleIds` are assigned values from the records instead from what has actually been revived. It turns out to be that even if nothing is revived, `usedChunkIds` and `usedModuleIds` will be equal to values from previous compilation taken from records – this is clearly not intended.

Because of this, chunk ids will change on every compilation if running in watch mode or using portable records.