Set correct exit code when question not answered
Apparently, questionInterface.question() never calls the answer callback
when the end of the (non-interactive) input stream is reached. To ensure
a correct exit status in this scenario we listen for the question interface
closing, check if the question was actually answered, and fall back to the
failure scenario if it was not.

Fixes #10076