Exit with an error if we can't prompt for webpack-cli installation.
If STDIN is not in terminal mode when the webpack script is executed, the
readline module will never execute the callback passed to question and the
script will exit successfully, without ever building the pack.  An example of
this is Rails' webpacker gem, which executes the webpack script in a subshell.

If STDIN is not in terminal mode, we should exit with an error immediately,
instead of attempting to ask a question to which we'll never receive an answer.