Remove isRequiredVersion utility function
In SharePlugin.js and ConsumeSharedPlugin.js, if a configuration value did not look like a version, it was silently ignored instead of generating an error. This is now changed to generate an error (i.e., we try to use it as a version) if would be interpreted as a version range, but is not.

Also, it appears that in ConsumeSharedPlugin.js, a configuration possibility accounted for in the code is to give webpack's internal version range encoding in the configuration (i.e., the array of version components as strings or numbers), so we also test for this possibility. Depending on how you look at it, this may make this encoding of the required version range a supported public API?

Finally, we change one more place in ConsumeSharedPlugin.js to actually parse the required version and throw an error if there is a problem, rather than attempting to check if it is valid first, which is more accurate.