Trace local file installs to guess the required version for a shared module.
Sometimes a module is installed with a local path instead of a version. If we do not recognize a version number, this checks to see if it is a path on disk. If it is, we get the required version from the source on disk.

Perhaps a better way to handle this situation is getting the actual installed version of the module. That would handle more exotic ways of installing, including from git, etc.