Improve suggested Node.js polyfill config
Resolve fallback to an absolute path in the suggested config for the
Node.js polyfill.  This is necessary in setups where where packages that
use the Node.js module are symlinked outside the package directory and
fail to resolve the fallback as a result.

A trailing slash is also necessary for packages that are shadowed by
builtin modules, i.e.: require.resolve("events") returns "events" while
require.resolve("events/") returns the path to the polyfill.