[#11670] Adds a warning about incorrect use of HMR chunks
Namely, this update adds to the beginning of each xxx.hot-update.js
chunk a simple check that the chunk is loaded by a <script> element with
"data-webpack" attribute set, which is expected when the chunk is loaded
by Webpack machinery. Failure of the check means that user loaded the
chunk himself by mistake, which would cause obscure errors within
Wepback, but this update does not let it happen, instead throwing an
error right away with a more clear message.