schemas: Allow empty array for dependOn, entry, import, noParse, paths
There may be no practical reason to use an empty array here, but it’s not
harmful, and allowing it makes the TypeScript typings more convenient,
e.g. string[] instead of [string, ...string[]].

The [string, ...string[]] type was annoying to work with because it’s
incompatible with almost every nontrivial way you might compute the value other
than writing it out literally. For example, using filter or map or importing
from a .json file all give you a string[], which cannot be assigned to [string,
...string[]] without an unchecked cast, even if the programmer “knows” the
result will be nonempty.

Fixes #11801.

Signed-off-by: Anders Kaseorg <andersk@mit.edu>