fix build cycle detection in Compilation
When tracking the modules which are being built by an origin module,
the code mistakenly adds the origin module itself to the set of modules
it's building, instead of the actual module about to be built. As a
result, when a module build finishes (and the module has itself called
loadModule()), it remains marked as building itself. So if it gets
loaded again, a BuildCycleError is reported despite no cycle existing.

Fixes #14379