Move types to devDependencies
I ran in to an issue with `@types/eslint` over here: https://github.com/typed-ember/ember-cli-typescript/issues/1502

here is how I came to conclude the issue was with webpack:
```
❯ pnpm why @types/eslint

devDependencies:
@embroider/webpack 1.6.0
├─┬ @embroider/babel-loader-8 1.6.0
│ └─┬ babel-loader 8.2.5
│   └─┬ webpack 5.72.0 peer
│     └─┬ @types/eslint-scope 3.7.3
│       └── @types/eslint 8.4.2
├─┬ @embroider/hbs-loader 1.6.0
│ └─┬ webpack 5.72.0 peer
│   └─┬ @types/eslint-scope 3.7.3
│     └── @types/eslint 8.4.2
├─┬ babel-loader 8.2.5
│ └─┬ webpack 5.72.0 peer
│   └─┬ @types/eslint-scope 3.7.3
│     └── @types/eslint 8.4.2
├─┬ css-loader 5.2.7
│ └─┬ webpack 5.72.0 peer
│   └─┬ @types/eslint-scope 3.7.3
│     └── @types/eslint 8.4.2

(etc, for a few more pages)
```

This is problematic, because I'm using eslint@v7. _but_ I'm not even working with eslint plugins in the project where I discovered this, and my type checking was throwing errors due to mismatched types versions:
```
node_modules/.pnpm/@types+eslint@7.29.0/node_modules/@types/eslint/helpers.d.ts:1:6 - error TS2300: Duplicate identifier 'Prepend'.

1 type Prepend<Tuple extends any[], Addend> = ((_: Addend, ..._1: Tuple) => any) extends (..._: infer Result) => any
       ~~~~~~~

  node_modules/.pnpm/@types+eslint@8.4.2/node_modules/@types/eslint/helpers.d.ts:1:6
    1 type Prepend<Tuple extends any[], Addend> = ((_: Addend, ..._1: Tuple) => any) extends (..._: infer Result) => any
           ~~~~~~~
    'Prepend' was also declared here.
```

This is, of course, solved by pinning / overrides / resolutions-ing to `@types/eslint @ v7, but this is a bandaid on the real problem -- that webpack is bringing in dev dependencies when it should not be.