Add `i64` to the set of JS-compatible wasm types in `syncWebAssembly` mode
For quite a while now, it's been possible for WebAssembly `i64`s to be converted to/from JS bigints (as function parameters, results, etc.). However, `syncWebAssembly` mode currently rejects any modules that attempt to do so, because `i64` isn't in it's list of JS-compatible types. This fixes that by adding `i64` to that list.

There was an existing test that used `i64` as an example of a non-JS-compatible type; I replaced that with `v128`.