Serialize code generator data to support generated assets
AssetGenerator calls out in a TODO-comment that the filename, assetInfo,
and fullContentHash values must be captured in the 'data' object that's
populated during code generation in order to be accessible in the
AssetModulesPlugin. It notes that it must store them in the code
generation results because it will be cached, but that appears to be
incorrect as data is a simple Map that's instantiated within the
NormalModule and not captured anywhere that would be cached.

As a result, configurations that use the asset/resource type and make
changes to assets between cached builds will result in a runtime error
as Webpack is able to access the file from cache but isn't able to
access the expected values from the data object.

This solution captures the data object as a property of the NormalModule
and hooks in to the existing serialize/deserialize functionality in
order to make this value available across cached builds.