Improve performance of hashRegExp lookup
For applications with a very large number of assets, the cost of
invoking a single regular expression with many many values in a group
becomes very high.

By changing to a list of regular expressions (with helper methods for
maintaining the original design) we can get a large performance
improvement.